fix: honor post-delivery disposition and ladder absence honestly
Map toggle/disclosure observation failures after a real write through
after_delivery, refuse to count ScrollPattern absence as ladder delivery, and
split oversized dispatch tests under the 400-line cap.
