+++ crates/windows/src/actions/mutation_tests.rs
@@ -217,6 +217,24 @@ fn write_path_sources_never_reach_the_read_classification_table() {
     );
 }
 
+#[test]
+fn a_planted_read_classification_call_is_caught() {
+    let banned = concat!("classify_read_", "hresult");
+    let fixture = "let _ = classify_read_hresult(hr);";
+    assert!(fixture.contains(banned));
+    let offences = fixture
+        .lines()
+        .filter(|line| {
+            let trimmed = line.trim_start();
+            !(trimmed.starts_with("///") || trimmed.starts_with("//!")) && line.contains(banned)
+        })
+        .count();
+    assert_eq!(
+        offences, 1,
+        "MUST-CATCH: planted classify_read_hresult must fail the write-path scan"
+    );
+}
+
 fn mutation_sources() -> [(&'static str, &'static str); 13] {
     [
         ("actions/mutation.rs", include_str!("mutation.rs")),
